Renamed invoker to core
