+++ core/entity/employee.hpp
@@ -18,44 +18,27 @@
 *           Ben Ziv <pointonsoftware@gmail.com>                                                   *
 *                                                                                                 *
 **************************************************************************************************/
-#ifndef INVOKER_ENTITY_PERSON_HPP_
-#define INVOKER_ENTITY_PERSON_HPP_
+#ifndef CORE_ENTITY_EMPLOYEE_HPP_
+#define CORE_ENTITY_EMPLOYEE_HPP_
 
 #include <string>
-#include <vector>
-#include "address.hpp"
-#include "contactdetails.hpp"
-#include "personalid.hpp"
+#include "person.hpp"
 
 namespace entity {
 
-class Person {
+class Employee : public Person {
  public:
-    Person(const std::string& firstname,
-           const std::string& middlename,
-           const std::string& lastname,
-           const std::string& birthdate,
-           const std::string& gender);
-    Person() = default;
-    virtual ~Person() = default;
-
-    void addPhoneNumber(const std::string& phonenumber);
-    void addPersonalId(const std::string& type, const std::string& number);
-    void setPersonalIds(const std::vector<PersonalId>& personalids);
-    void setEmail(const std::string& email);
-    void setAddress(const Address& address);
+    Employee(const std::string& firstname,
+             const std::string& middlename,
+             const std::string& lastname,
+             const std::string& birthdate,
+             const std::string& gender);
+    Employee() = default;
+    ~Employee() = default;
 
  private:
-    std::string m_firstname;
-    std::string m_middlename;
-    std::string m_lastname;
-    std::string m_birthdate;
-    std::string m_gender;
-    std::string m_email;
-    Address m_address;
-    ContactDetails m_contact_details;
-    std::vector<PersonalId> m_personal_ids;
+    std::string position;
 };
 
 }  // namespace entity
-#endif  // INVOKER_ENTITY_PERSON_HPP_
+#endif  // CORE_ENTITY_EMPLOYEE_HPP_
