--- invoker/CMakeLists.txt
@@ -1,4 +1,0 @@
-project (invoker)
-
-
-add_subdirectory (entity)
